Add system test
System test added in the form of a bash script. It tests if the app reads from the
standard input and prints to the standard output. It also tests the command line
arguments.

This would be much trickier to test from C++, as there is no standard way of capturing
the stdin and stdout.

This is also not standard, as it requires bash that is not available on every OS.
(E.g. even Alpine Linux does not have it by default.)